Updated SingleListingResponseResult.java. We can now enter the individual listing again!

+++ frontend/app/src/main/java/com/chads/vanroomies/SingleListingResponseResult.java
@@ -108,23 +108,23 @@ public class SingleListingResponseResult {
     }
 
     public static class LocationObj {
-        private int latitude;
-        private int longitude;
+        private Double latitude;
+        private Double longitude;
         private String _id;
 
-        public int getLatitude() {
+        public Double getLatitude() {
             return latitude;
         }
 
-        public void setLatitude(int latitude) {
+        public void setLatitude(Double latitude) {
             this.latitude = latitude;
         }
 
-        public int getLongitude() {
+        public Double getLongitude() {
             return longitude;
         }
 
-        public void setLongitude(int longitude) {
+        public void setLongitude(Double longitude) {
             this.longitude = longitude;
         }
